soundmind-ai-desk/backend/requirements.txt
Flask
flask-cors
python-dotenv
openai
mysql-connector-python
sqlalchemy
whisper
torch
ffmpeg-python
librosa
scikit-learn
